requirements.txt
polars>=0.20.0
openpyxl>=3.1.0
xlrd>=2.0.0
pyyaml>=6.0

